Fixes related to Review of PR #https://github.com/Cartopol/STS_Community_Warlord_Mod/pull/10
Please see the following PR for the changes made. https://github.com/Cartopol/STS_Community_Warlord_Mod/pull/10
