+++ src/main/java/the_warlord/cards/warlord/FocusPunch.java
@@ -1,16 +1,11 @@
 package the_warlord.cards.warlord;
 
-import com.megacrit.cardcrawl.actions.AbstractGameAction;
 import com.megacrit.cardcrawl.actions.common.ApplyPowerAction;
 import com.megacrit.cardcrawl.characters.AbstractPlayer;
-import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 import com.megacrit.cardcrawl.monsters.AbstractMonster;
 import the_warlord.WarlordMod;
-import the_warlord.actions.DoubleGushAction;
-import the_warlord.actions.TripleGushAction;
 import the_warlord.cards.CustomWarlordModCard;
 import the_warlord.characters.Warlord;
-import the_warlord.powers.BoobyTrapPower;
 import the_warlord.powers.FocusPunchPower;
 
 public class FocusPunch extends CustomWarlordModCard {
